+++ Cargo.toml
@@ -7,8 +7,8 @@ license = "MIT"
 [dependencies]
 http-body-util = "0.1.3"
 hyper = { version = "1.7.0", features = ["http1", "http2"] }
-hyper-util = { version = "0.1.16", features = ["server-auto", "tokio"] }
-serde = { version = "1.0.219", features = ["derive"] }
+hyper-util = { version = "0.1.17", features = ["server-auto", "tokio"] }
+serde = { version = "1.0.225", features = ["derive"] }
 tokio = { version = "1.47.1", features = ["rt-multi-thread", "macros", "signal"] }
 tracing = "0.1.41"
 tracing-subscriber = { version = "0.3.20", features = ["json", "env-filter"] }
@@ -16,10 +16,10 @@ uuid = { version = "1.18.1", features = ["v4"] }
 rustls-pemfile = "2.2.0"
 rustls-pki-types = "1.12.0"
 rustls = "0.23.31"
-tokio-rustls = "0.26.2"
+tokio-rustls = "0.26.3"
 async-trait = "0.1.89"
 reqwest = { version = "0.12.23", default-features = false, features = ["rustls-tls"] }
-config = { version = "0.15.15", default-features = false, features = ["yaml"] }
+config = { version = "0.15.16", default-features = false, features = ["yaml"] }
 thiserror = "2.0.16"
 humantime-serde = "1.1.1"
 tokio-util = "0.7.16"
@@ -30,3 +30,4 @@ arc-swap = "1.7.1"
 [profile.release]
 codegen-units = 1
 lto = true
+debug = true
